More fixes to popmap loading + move popmap log before function started log

+++ src/freq.cpp
@@ -13,7 +13,7 @@ void freq(Parameters& parameters) {
 
     Header header = get_header(parameters.markers_table_path);
     Popmap popmap;  // Create dummy popmap
-    uint n_individuals = header.size() - 1; // Number of columns - 2 (id and seq columns) +1 (because range is 0 - n_individuals)
+    uint n_individuals = static_cast<uint>(header.size()) - 1; // Number of columns - 2 (id and seq columns) +1 (because range is 0 - n_individuals)
 
     bool parsing_ended = false;
     MarkersQueue markers_queue;
